+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Add upgrade considerations for Wagtail 5.1
+
 ## 1.0.0 (2023-07-25)
 
 - Integrate [wagtail upgrades](https://github.com/unexceptable/wagtail-robots/pull/20)